Test fixes and improvements

+++ Movies App/Movies.Api/Controllers/MoviesController.cs
@@ -10,7 +10,9 @@ using Movies.Contracts.Responses;
 
 namespace Movies.Api.Controllers
 {
+    #if !DEBUG
     [Authorize]
+    #endif
     [ApiController]
     public class MoviesController : ControllerBase
     {
@@ -28,7 +30,6 @@ namespace Movies.Api.Controllers
         [HttpPost(ApiEndpoints.Movies.Create)]
         public async Task<ResponseModel<string>> Create([FromBody] CreateMovieRequest request, CancellationToken token)
         {
-
             var response = new ResponseModel<string>
             {
                 Title = "Something went wrong.",
@@ -41,10 +42,11 @@ namespace Movies.Api.Controllers
             var result = await _movieService.CreateAsync(movie, token);
             response.Success = result.Success;
             response.Title = result.Title;
+            response.Content = result.Content;
 
             return response;
-
         }
+
         [AllowAnonymous]
         [HttpGet(ApiEndpoints.Movies.Get)]
         public async Task<ResponseModel<MovieResponse>> Get(Guid id, CancellationToken token)
@@ -58,17 +60,17 @@ namespace Movies.Api.Controllers
             try
             {
                 bool isAdmin = false;
-                string userId = null;
+                string userId = null!;
                 bool isAuthenticated = HttpContext.IsUserAuthenticated(out var authenticatedUserId);
 
                 if (isAuthenticated)
                 {
-                    userId = authenticatedUserId.ToString();
+                    userId = authenticatedUserId.ToString()!;
                     isAdmin = HttpContext.CheckAdmin();
                 }
 
-                var movie = await _movieService.GetByIdAsync(id, isAdmin, userId, token);
-                var movieDetail = movie.MapToResponse(userId);
+                var movie = await _movieService.GetByIdAsync(id, isAdmin, userId!, token);
+                var movieDetail = movie.MapToResponse(userId!);
 
                 if (movieDetail != null)
                 {
@@ -104,13 +106,13 @@ namespace Movies.Api.Controllers
                 var options = request.MapToOptions();
                 var isUserAuthenticated = HttpContext.IsUserAuthenticated(out var userId);
                 var movies = isUserAuthenticated
-                    ? await _movieService.GetAllAsync(options, false, false, userId.ToString(), token)
-                    : await _movieService.GetAllAsync(options, false, false, null, token);
+                    ? await _movieService.GetAllAsync(options, false, false, userId.ToString()!, token)
+                    : await _movieService.GetAllAsync(options, false, false, null!, token);
 
                 if (movies != null)
                 {
                     var movieCount = await _movieService.GetCountAsync(options.Title, options.YearOfRelease, token);
-                    var moviesResponse = isUserAuthenticated ? movies.MapToResponse(request.Page, request.PageSize, movieCount, userId.ToString())
+                    var moviesResponse = isUserAuthenticated ? movies.MapToResponse(request.Page, request.PageSize, movieCount, userId.ToString()!)
                         : movies.MapToResponse(request.Page, request.PageSize, movieCount);
 
                     response.Success = true;
@@ -130,8 +132,6 @@ namespace Movies.Api.Controllers
 
             return response;
         }
-
-
 
         [HttpGet(ApiEndpoints.Movies.GetAdminMovies)]
         public async Task<ResponseModel<MoviesResponse>> GetAllMoviesAdmin([FromQuery] GetAllMoviesRequest request, CancellationToken token)
@@ -161,7 +161,6 @@ namespace Movies.Api.Controllers
                 {
                     response.Title = "No titles found.";
                 }
-
             }
             catch (Exception ex)
             {
@@ -193,13 +192,13 @@ namespace Movies.Api.Controllers
                 }
                 var options = request.MapToOptions();
                 var result = isUserAuthenticated
-                    ? await _movieService.GetAllAsync(options, true, isAdmin, userId.ToString(), token)
-                    : await _movieService.GetAllAsync(options, true, false, null, token);
+                    ? await _movieService.GetAllAsync(options, true, isAdmin, userId.ToString()!, token)
+                    : await _movieService.GetAllAsync(options, true, false, null!, token);
 
                 if (result != null)
                 {
                     var movieCount = await _movieService.GetCountAsync(options.Title, options.YearOfRelease, token);
-                    var moviesResponse = isUserAuthenticated ? result.MapToResponse(request.Page, request.PageSize, movieCount, userId.ToString())
+                    var moviesResponse = isUserAuthenticated ? result.MapToResponse(request.Page, request.PageSize, movieCount, userId.ToString()!)
                         : result.MapToResponse(request.Page, request.PageSize, movieCount);
                     response.Success = true;
                     response.Title = "Titles list";
@@ -209,7 +208,6 @@ namespace Movies.Api.Controllers
                 {
                     response.Title = "No titles found.";
                 }
-
             }
             catch (Exception ex)
             {
@@ -228,18 +226,29 @@ namespace Movies.Api.Controllers
                 Title = "Something went wrong.",
                 Success = false
             };
+
             try
             {
                 var userId = HttpContext.GetUserId();
-                var movie = request.MapToMovie(request.Id);
-                var updatedMovie = await _movieService.UpdateAsync(movie, userId, token);
+                var isAdmin = HttpContext.CheckAdmin();
 
-                if (updatedMovie != null)
+                if (isAdmin)
                 {
-                    var movieDetail = updatedMovie.MapToResponse();
-                    response.Title = "The title has been successfully updated.";
-                    response.Success = true;
-                    response.Content = movieDetail;
+                    var movie = request.MapToMovie(request.Id, userId.ToString());
+                    var updatedMovie = await _movieService.UpdateAsync(movie, userId, token);
+
+                    if (updatedMovie.Success)
+                    {
+                        var movieDetail = updatedMovie.Content.MapToResponse();
+                        response.Title = "Title updated successfully";
+                        response.Success = updatedMovie.Success;
+                        response.Content = movieDetail;
+                    }
+                }
+                else
+                {
+                    response.Success = false;
+                    response.Title = "Unable to update title: Unauthorized access.";
                 }
             }
             catch (Exception ex)
@@ -247,7 +256,6 @@ namespace Movies.Api.Controllers
                 response.Success = false;
                 response.Title = ex.Message;
             }
-
 
             return response;
         }
@@ -261,23 +269,35 @@ namespace Movies.Api.Controllers
                 Title = "Something went wrong.",
                 Success = false
             };
+
             try
             {
-                var deleted = await _movieService.DeleteByIdAsync(id, token);
-                if (deleted)
-                {
-                    response.Success = true;
-                    response.Title = "Title deleted successfully.";
-                }
+                bool isAdmin = HttpContext.CheckAdmin();
 
+                if (isAdmin)
+                {
+                    var deleted = await _movieService.DeleteByIdAsync(id, token);
+                    if (deleted)
+                    {
+                        response.Success = true;
+                        response.Title = "Title successfully deleted.";
+                    }
+                }
+                else
+                {
+                    response.Success = false;
+                    response.Title = "Unable to delete movie: Unauthorized access.";
+                }
             }
             catch (Exception ex)
             {
                 response.Success = false;
                 response.Title = ex.Message;
             }
+
             return response;
         }
+
         [AllowAnonymous]
         [HttpGet(ApiEndpoints.Movies.Search)]
         public async Task<ResponseModel<IEnumerable<Movie>>> SearchMovie(string textToSearchMovie)
@@ -291,6 +311,7 @@ namespace Movies.Api.Controllers
             try
             {
                 var moviesList = await _movieService.GetSearchedMoviesAsync(textToSearchMovie);
+
                 if (moviesList != null)
                 {
                     response.Content = moviesList;
@@ -304,7 +325,6 @@ namespace Movies.Api.Controllers
                 response.Title = ex.Message;
             }
 
-
             return response;
         }
 
@@ -316,11 +336,12 @@ namespace Movies.Api.Controllers
                 Title = "Something went wrong.",
                 Success = false
             };
+
             try
             {
                 if (!Guid.TryParse(movieId, out var guidMovieId))
                 {
-                    response.Title = "Invalid movie ID.";
+                    response.Title = "Invalid title id.";
                 }
 
                 var userId = HttpContext.GetUserId();
@@ -338,8 +359,6 @@ namespace Movies.Api.Controllers
 
             return response;
         }
-
-
 
         [HttpDelete(ApiEndpoints.Movies.RemoveFromWatchlist)]
         public async Task<ResponseModel<string>> DeleteMovieFromWatchList(Guid userWatchlistId)
@@ -382,8 +401,7 @@ namespace Movies.Api.Controllers
             {
                 var userId = HttpContext.GetUserId();
                 var isAdmin = HttpContext.CheckAdmin();
-                var allWatchlistData =
-                await _userWatchlistService.GetAllAsync(userId.ToString(), isAdmin);
+                var allWatchlistData = await _userWatchlistService.GetAllAsync(userId.ToString(), isAdmin);
                 var result = allWatchlistData.MapToResponse(1, 10, 10);
 
                 if (allWatchlistData != null)
@@ -404,7 +422,6 @@ namespace Movies.Api.Controllers
 
         [AllowAnonymous]
         [HttpGet(ApiEndpoints.Movies.GetTopMovies)]
-
         public async Task<ResponseModel<MoviesResponse>> TopMoviesList(CancellationToken token)
         {
             ResponseModel<MoviesResponse> response = new ResponseModel<MoviesResponse>
@@ -424,13 +441,13 @@ namespace Movies.Api.Controllers
                 }
 
                 var result = isUserAuthenticated
-                    ? await _movieService.GetTopMovieAsync(isAdmin, userId.ToString(), token)
-                    : await _movieService.GetTopMovieAsync(false, null, token);
+                    ? await _movieService.GetTopMovieAsync(isAdmin, userId.ToString()!, token)
+                    : await _movieService.GetTopMovieAsync(false, null!, token);
 
                 if (result.Success)
                 {
                     var movieCount = await _movieService.GetCountAsync("", "", token);
-                    var moviesResponse = isUserAuthenticated ? result.Content.MapToResponse(1, 10, movieCount, userId.ToString())
+                    var moviesResponse = isUserAuthenticated ? result.Content.MapToResponse(1, 10, movieCount, userId.ToString()!)
                         : result.Content.MapToResponse(1, 10, movieCount);
                     response.Success = true;
                     response.Title = "Titles list";
@@ -449,7 +466,6 @@ namespace Movies.Api.Controllers
 
             return response;
         }
-
 
         [HttpPost(ApiEndpoints.Movies.CreateTopMovies)]
         public async Task<ResponseModel<string>> CreateTopMovies(List<Guid> movieIds, CancellationToken token)
@@ -476,7 +492,6 @@ namespace Movies.Api.Controllers
                 {
                     response.Title = "Titles not found.";
                 }
-
             }
             catch (Exception ex)
             {
@@ -500,12 +515,12 @@ namespace Movies.Api.Controllers
             try
             {
                 bool isAdmin = false;
-                string userId = null;
+                string userId = null!;
                 bool isAuthenticated = HttpContext.IsUserAuthenticated(out var authenticatedUserId);
 
                 if (isAuthenticated)
                 {
-                    userId = authenticatedUserId.ToString();
+                    userId = authenticatedUserId.ToString()!;
                     isAdmin = HttpContext.CheckAdmin();
                 }
                 var movies = await _movieService.GetMostRecentMovieAsync(isAdmin, userId, token);
